# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "XSecurity-Vault",
    "description": "Secure digital certificate management orchestrator",
    "version": "2.0.0"
  },
  "paths": {
...
    "/usuarios/": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "List Users",
        "description": "List users in organization.\n\nIf organization_id is not provided, uses the current user's organization.",
        "operationId": "list_users_usuarios__get",
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Organization Id"
            }
          },
          {
            "name": "include_deleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Deleted"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{user_id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Get User",
        "description": "Get specific user by ID.",
        "operationId": "get_user_usuarios__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get User Usuarios  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Update User",
        "description": "Update user.\n\nOnly organization owners can update users.",
        "operationId": "update_user_usuarios__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update User Usuarios  User Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Delete User",
        "description": "Soft delete user.\n\nOnly organization owners can delete users.",
        "operationId": "delete_user_usuarios__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete User Usuarios  User Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/grupos/{grupo_id}": {
      "get": {
        "tags": [
          "Grupos"
        ],
        "summary": "Obter",
        "operationId": "obter_grupos__grupo_id__get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Grupos"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_grupos__grupo_id__put",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Grupos"
        ],
        "summary": "Remover",
        "operationId": "remover_grupos__grupo_id__delete",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/grupos/": {
      "post": {
        "tags": [
          "Grupos"
        ],
        "summary": "Criar",
        "operationId": "criar_grupos__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos/{grupo_id}/usuarios": {
      "get": {
        "tags": [
          "Grupos"
        ],
        "summary": "Listar Usuarios Do Grupo",
        "operationId": "listar_usuarios_do_grupo_grupos__grupo_id__usuarios_get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Grupos"
        ],
        "summary": "Adicionar Usuario Ao Grupo",
        "operationId": "adicionar_usuario_ao_grupo_grupos__grupo_id__usuarios_post",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos/{grupo_id}/usuarios/{usuario_id}": {
      "delete": {
        "tags": [
          "Grupos"
        ],
        "summary": "Remover Usuario Do Grupo",
        "operationId": "remover_usuario_do_grupo_grupos__grupo_id__usuarios__usuario_id__delete",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          },
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Usuario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos/{grupo_id}/usuarios/bulk": {
      "post": {
        "tags": [
          "Grupos"
        ],
        "summary": "Adicionar Usuarios Em Lote",
        "operationId": "adicionar_usuarios_em_lote_grupos__grupo_id__usuarios_bulk_post",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrupoUsuarioBulkCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/grupos-certificados/grupo/{grupo_id}": {
      "get": {
        "tags": [
          "Grupos Certificados"
        ],
        "summary": "Listar Grupo",
        "operationId": "listar_grupo_grupos_certificados_grupo__grupo_id__get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GrupoCertOut"
                  },
                  "title": "Response Listar Grupo Grupos Certificados Grupo  Grupo Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/grupos-usuarios/": {
      "get": {
        "tags": [
          "Grupos Usu\u00e1rios"
        ],
        "summary": "Listar",
        "operationId": "listar_grupos_usuarios__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GrupoUsuarioOut"
                  },
                  "type": "array",
                  "title": "Response Listar Grupos Usuarios  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Grupos Usu\u00e1rios"
        ],
        "summary": "Criar",
        "operationId": "criar_grupos_usuarios__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrupoUsuarioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrupoUsuarioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos-usuarios/grupo/{grupo_id}": {
      "get": {
        "tags": [
          "Grupos Usu\u00e1rios"
        ],
        "summary": "Listar Grupo",
        "operationId": "listar_grupo_grupos_usuarios_grupo__grupo_id__get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GrupoUsuarioOut"
                  },
                  "title": "Response Listar Grupo Grupos Usuarios Grupo  Grupo Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos-usuarios/{grupo_usuario_id}": {
      "get": {
        "tags": [
          "Grupos Usu\u00e1rios"
        ],
        "summary": "Obter",
        "operationId": "obter_grupos_usuarios__grupo_usuario_id__get",
        "parameters": [
          {
            "name": "grupo_usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Usuario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrupoUsuarioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Grupos Usu\u00e1rios"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_grupos_usuarios__grupo_usuario_id__put",
        "parameters": [
          {
            "name": "grupo_usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Usuario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrupoUsuarioUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrupoUsuarioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Grupos Usu\u00e1rios"
        ],
        "summary": "Deletar",
        "operationId": "deletar_grupos_usuarios__grupo_usuario_id__delete",
        "parameters": [
          {
            "name": "grupo_usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Usuario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos-usuarios/bulk": {
      "post": {
        "tags": [
          "Grupos Usu\u00e1rios"
        ],
        "summary": "Criar Bulk",
        "description": "Cria v\u00ednculos em lote entre um grupo e m\u00faltiplos usu\u00e1rios. Retorna os ids criados e os pulados com raz\u00e3o.",
        "operationId": "criar_bulk_grupos_usuarios_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrupoUsuarioBulkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrupoUsuarioBulkResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/regras-acesso/grupo/{grupo_id}": {
      "get": {
        "tags": [
          "Regras de Acesso"
        ],
        "summary": "Listar Grupo",
        "operationId": "listar_grupo_regras_acesso_grupo__grupo_id__get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegraAcessoOut"
                  },
                  "title": "Response Listar Grupo Regras Acesso Grupo  Grupo Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/regras-acesso-hosts/grupo/{grupo_id}": {
      "get": {
        "tags": [
          "Regras de Acesso (Hosts)"
        ],
        "summary": "Listar Grupo",
        "operationId": "listar_grupo_regras_acesso_hosts_grupo__grupo_id__get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegraAcessoHostOut"
                  },
                  "title": "Response Listar Grupo Regras Acesso Hosts Grupo  Grupo Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GrupoCertOut": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "grupo_cert_id": {
            "type": "integer",
            "title": "Grupo Cert Id"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "grupo_cert_id"
        ],
        "title": "GrupoCertOut"
      },
...
      "GrupoUsuarioBulkCreate": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "usuario_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Usuario Ids"
          },
          "empresa_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Empresa Id"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "usuario_ids"
        ],
        "title": "GrupoUsuarioBulkCreate"
      },
      "GrupoUsuarioBulkResult": {
        "properties": {
          "created": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Created"
          },
          "skipped": {
            "items": {
              "$ref": "#/components/schemas/GrupoUsuarioBulkSkipped"
            },
            "type": "array",
            "title": "Skipped"
          }
        },
        "type": "object",
        "required": [
          "created",
          "skipped"
        ],
        "title": "GrupoUsuarioBulkResult"
      },
      "GrupoUsuarioBulkSkipped": {
        "properties": {
          "usuario_id": {
            "type": "integer",
            "title": "Usuario Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "usuario_id",
          "reason"
        ],
        "title": "GrupoUsuarioBulkSkipped"
      },
      "GrupoUsuarioCreate": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          }
        },
        "type": "object",
        "required": [
          "grupo_id"
        ],
        "title": "GrupoUsuarioCreate"
      },
      "GrupoUsuarioOut": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "grupo_usuario_id": {
            "type": "integer",
            "title": "Grupo Usuario Id"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "grupo_usuario_id"
        ],
        "title": "GrupoUsuarioOut"
      },
      "GrupoUsuarioUpdate": {
        "properties": {
          "grupo_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grupo Id"
          }
        },
        "type": "object",
        "title": "GrupoUsuarioUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse",
        "description": "Paginated response wrapper."
      },
...
      "RegraAcessoOut": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "tipo_dia": {
            "$ref": "#/components/schemas/TipoDiaEnum"
          },
          "dias_especificos": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Especificos"
          },
          "horarios": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Horarios"
          },
          "regra_id": {
            "type": "integer",
            "title": "Regra Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "tipo_dia",
          "horarios",
          "regra_id",
          "criado_em"
        ],
        "title": "RegraAcessoOut"
      },
...
      "UserUpdate": {
        "properties": {
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "requires_2fa": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requires 2Fa"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "User update request."
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "Session token for authentication. Use the session_token from login response.",
        "scheme": "bearer"
      }
    }
  }
}